refactor(react-utils): rename useKeyedAsyncValue to useAsyncMemo
