+++ packages/react-utils/src/index.ts
@@ -1,7 +1,7 @@
 export { useDebounce } from './hooks/useDebounce';
 export { useDebouncedValue } from './hooks/useDebouncedValue';
 export { useDidUpdate } from './hooks/useDidUpdate';
-export { useKeyedAsyncValue } from './hooks/useKeyedAsyncValue';
+export { useKeyedAsyncValue } from './hooks/useAsyncMemo';
 export { useKeyPress } from './hooks/useKeyPress';
 export { useOnce } from './hooks/useOnce';
 export { useOnClickOutside } from './hooks/useOnClickOutside';
